third_party/blink/web_tests/fast/css/fontfaceset-events-expected.txt

Test that the event handlers of FontFaceSet are called in the correct order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.type is "loading"
PASS event.type is "loadingdone"
PASS numberOfEvents is 2
PASS successfullyParsed is true

TEST COMPLETE

Hello, world!
